djaysaday/requirements.txt
Django~=2.2.4
djangorestframework
markdown
django-filter